split-to-file-byqr/requirements.txt
requests==2.27.1
opencv-python==4.5.5.64
Flask==2.1.2
Pillow==9.1.1
pdf2image==1.16.0
